requirements.txt
pyserial
xmltodict
pyperclip
PyQt6
